+++ apiary-metastore-events/apiary-hive-events/src/main/java/com/expediagroup/apiary/extensions/events/metastore/event/ApiaryInsertEvent.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2019 Expedia, Inc.
+ * Copyright (C) 2018-2020 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,10 +15,14 @@
  */
 package com.expediagroup.apiary.extensions.events.metastore.event;
 
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 
+import org.apache.hadoop.hive.metastore.api.FieldSchema;
+import org.apache.hadoop.hive.metastore.api.Partition;
+import org.apache.hadoop.hive.metastore.api.Table;
 import org.apache.hadoop.hive.metastore.events.InsertEvent;
 
 public class ApiaryInsertEvent extends ApiaryListenerEvent {
@@ -34,11 +38,25 @@ public class ApiaryInsertEvent extends ApiaryListenerEvent {
 
   public ApiaryInsertEvent(InsertEvent event) {
     super(event);
-    databaseName = event.getDb();
-    tableName = event.getTable();
-    partitionKeyValues = event.getPartitionKeyValues();
+
+    Table table = event.getTableObj();
+    databaseName = table.getDbName();
+    tableName = table.getTableName();
+    partitionKeyValues = createPartitionKeyValues(table, event.getPartitionObj());
     files = event.getFiles();
     fileChecksums = event.getFileChecksums();
+  }
+
+  public static LinkedHashMap<String, String> createPartitionKeyValues(Table table, Partition partition) {
+    LinkedHashMap<String, String> partitionKeyValues = new LinkedHashMap<>();
+    if (partition != null) {
+      List<FieldSchema> partitionKeys = table.getPartitionKeys();
+      List<String> partitionValues = partition.getValues();
+      for (int i = 0; i < partitionKeys.size(); i++) {
+        partitionKeyValues.put(partitionKeys.get(i).getName(), partitionValues.get(i));
+      }
+    }
+    return partitionKeyValues;
   }
 
   @Override